requirements.txt
# Requirements for Distributed AI Training System

# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Kafka
kafka-python==2.0.2

# MongoDB
pymongo==4.6.1

# Configuration
pyyaml==6.0.1

# System Monitoring
psutil==5.9.7

# Image Processing
pillow==10.1.0

# Data Processing
numpy==1.26.2
pandas==2.1.4

# HTTP Client
requests==2.31.0

# Logging
python-json-logger==2.0.7

# Development Dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
